backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg==3.2.3
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.17
websockets==14.1
pyyaml==6.0.2
python-dateutil==2.9.0.post0
